+++ src/solver/BeamSolver.cpp
@@ -179,7 +179,7 @@ void CBeamSolver::execute()
         for (j=0; j<6; j++)
             DofTopo(j+7) = beam->getNode(1)->getDof(j)->getNumber();
 
-        bandwidth = max(DofTopo) - min(DofTopo);
+        bandwidth = (int)max(DofTopo) - (int)min(DofTopo);
         if (bandwidth>maxBandwidth)
             maxBandwidth = bandwidth;
     }
@@ -269,7 +269,7 @@ void CBeamSolver::execute()
         nodeLoad->getDirection(vx, vy, vz);
         value = nodeLoad->getValue();
 
-        for (j=0; j<nodeLoad->getNodeSize(); j++)
+        for (j=0; j<(int)nodeLoad->getNodeSize(); j++)
         {
             CFemNode* node = nodeLoad->getNode(j);
             m_f(node->getDof(0)->getNumber()) = vx*value;
@@ -474,7 +474,7 @@ void CBeamSolver::execute()
     // Store element forces in elements
     //
 
-    double n;
+    int n;
     RowVector Ed(12);
     Matrix Es;
     Matrix Edi;
@@ -708,7 +708,7 @@ void CBeamSolver::update()
     // Store element forces in elements
     //
 
-    double n;
+    int n;
     RowVector Ed(12);
     Matrix Es;
     Matrix Edi;
